tycker: split `MockTycker` with `LetListTycker`
