+++ base/src/main/java/org/aya/tyck/tycker/UnifiedTycker.java
@@ -44,7 +44,7 @@ import java.util.function.Function;
  * @see #checkBoundaries
  * @see #inheritFallbackUnify
  */
-public sealed abstract class UnifiedTycker extends MockedTycker permits PropTycker {
+public sealed abstract class UnifiedTycker extends LetListTycker permits PropTycker {
   protected UnifiedTycker(@NotNull Reporter reporter, Trace.@Nullable Builder traceBuilder, @NotNull TyckState state) {
     super(reporter, traceBuilder, state);
   }
